Upload assets before item tests
